+++ sei-ibc-go/testing/mock/ibc_module.go
@@ -2,6 +2,7 @@ package mock
 
 import (
 	"bytes"
+	"fmt"
 	"strconv"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -52,8 +53,8 @@ func (im IBCModule) OnChanOpenTry(
 ) (version string, err error) {
 	if im.IBCApp.OnChanOpenTry != nil {
 		return im.IBCApp.OnChanOpenTry(ctx, order, connectionHops, portID, channelID, chanCap, counterparty, counterpartyVersion)
-
 	}
+
 	// Claim channel capability passed back by IBC module
 	if err := im.scopedKeeper.ClaimCapability(ctx, chanCap, host.ChannelCapabilityPath(portID, channelID)); err != nil {
 		return "", err
@@ -105,12 +106,13 @@ func (im IBCModule) OnRecvPacket(ctx sdk.Context, packet channeltypes.Packet, re
 	}
 
 	// set state by claiming capability to check if revert happens return
-	_, err := im.scopedKeeper.NewCapability(ctx, MockRecvCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
-	if err != nil {
+	capName := GetMockRecvCanaryCapabilityName(packet)
+	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)
 	}
+
 	if bytes.Equal(MockPacketData, packet.GetData()) {
 		return MockAcknowledgement
 	} else if bytes.Equal(MockAsyncPacketData, packet.GetData()) {
@@ -126,8 +128,8 @@ func (im IBCModule) OnAcknowledgementPacket(ctx sdk.Context, packet channeltypes
 		return im.IBCApp.OnAcknowledgementPacket(ctx, packet, acknowledgement, relayer)
 	}
 
-	_, err := im.scopedKeeper.NewCapability(ctx, MockAckCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
-	if err != nil {
+	capName := GetMockAckCanaryCapabilityName(packet)
+	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)
@@ -142,12 +144,27 @@ func (im IBCModule) OnTimeoutPacket(ctx sdk.Context, packet channeltypes.Packet,
 		return im.IBCApp.OnTimeoutPacket(ctx, packet, relayer)
 	}
 
-	_, err := im.scopedKeeper.NewCapability(ctx, MockTimeoutCanaryCapabilityName+strconv.Itoa(int(packet.GetSequence())))
-	if err != nil {
+	capName := GetMockTimeoutCanaryCapabilityName(packet)
+	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)
 	}
 
 	return nil
+}
+
+// GetMockRecvCanaryCapabilityName generates a capability name for testing OnRecvPacket functionality.
+func GetMockRecvCanaryCapabilityName(packet channeltypes.Packet) string {
+	return fmt.Sprintf("%s%s%s%s", MockRecvCanaryCapabilityName, packet.GetDestPort(), packet.GetDestChannel(), strconv.Itoa(int(packet.GetSequence())))
+}
+
+// GetMockAckCanaryCapabilityName generates a capability name for OnAcknowledgementPacket functionality.
+func GetMockAckCanaryCapabilityName(packet channeltypes.Packet) string {
+	return fmt.Sprintf("%s%s%s%s", MockAckCanaryCapabilityName, packet.GetSourcePort(), packet.GetSourceChannel(), strconv.Itoa(int(packet.GetSequence())))
+}
+
+// GetMockTimeoutCanaryCapabilityName generates a capability name for OnTimeoutacket functionality.
+func GetMockTimeoutCanaryCapabilityName(packet channeltypes.Packet) string {
+	return fmt.Sprintf("%s%s%s%s", MockTimeoutCanaryCapabilityName, packet.GetSourcePort(), packet.GetSourceChannel(), strconv.Itoa(int(packet.GetSequence())))
 }